[Pytorch Mobile] Optimize Non Forward for Mobile
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53314

Introduction of api for optimizing non forward functions for mobile. As of this diff, all functions that you say to optimize will be preserved, and those functions will be run through canonical optimization. The intention is to stack each further optimization onto separate diffs since they touch multiple files, and it seems like it'd be a nightmare to review.
ghstack-source-id: 123572503

Differential Revision: [D26618689](https://our.internmc.facebook.com/intern/diff/D26618689/)